pyproject.toml
[tool.poetry]
name = "apollo"
version = "0.1.0"
description = ""
authors = ["Veirt"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
